fix(receiver): type

+++ datagrump/receiver.cc
@@ -36,7 +36,7 @@ int main( int argc, char *argv[] )
   /* counter for the number of received datagrams for Delayed ACK */
   int seq_counter = 0;
 
-  int should_ack_num = 0;
+  uint64_t should_ack_num = 0;
 
   /* Loop and acknowledge every incoming datagram back to its source */
   while ( true ) {
